changed Decimal to float

+++ app/models/service.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from typing import List, Optional
 from enum import Enum
 from pydantic import BaseModel, Field
@@ -45,7 +44,7 @@ class Car(BaseModel):
 
 class Parts(Document):
     name: str
-    price: Decimal
+    price: float
     models: List[Car]
 
     class Settings:
